break content out from main dao
