+++ ExtendedLearning/src/main/java/com/skilldistillery/sdelp/data/ContentDAO.java
@@ -1,0 +1,11 @@
+package com.skilldistillery.sdelp.data;
+
+import com.skilldistillery.sdelp.entities.Content;
+
+public interface ContentDAO {
+	
+	Content addContent(Content content);
+	
+	Content updateContent(Content content);
+
+}
